requirements-dev.txt
pytest
sphinx
sphinx-gallery
sphinx_rtd_theme
numpydoc
matplotlib
mlxtend